feat: remove simulator restrictions for Python FBs and server/remote-device creation
Extract duplicated Python + C++ POU preprocessing into a shared
preprocessPous() utility. For the simulator target, Python function
blocks compile as empty ST stubs instead of generating the full
Python runtime wrapper (which requires POSIX shared memory unavailable
on AVR). Unlock server and remote-device creation UI for the simulator
target alongside OpenPLC Runtime v4, and suppress the board-switch
warning modal when switching to the simulator with existing servers
or remote devices.
